Image Classifier/requirements.txt
uvicorn
numpy==1.19.1
Pillow
pathlib2
requests
matplotlib 
tensorflow==2.7.0  
sklearn 
fastapi
  
